+++ Automation/Conditions/DroughtStartedCondition.cs
@@ -27,8 +27,10 @@ public sealed class DroughtStartedCondition : WeatherTrackerConditionBase {
   }
 
   /// <inheritdoc/>
-  protected override void OnWeatherChanged(bool isDrought) {
-    ConditionState = isDrought;
+  protected override void OnWeatherChanged(bool? isDrought, bool? isBadtide) {
+    if (isDrought.HasValue) {
+      ConditionState = isDrought.Value;
+    }
   }
   #endregion
 }