+++ Automation/package/CHANGES.md
@@ -1,5 +1,6 @@
 # v0.13 (pre-relealse):
 * Add "finish now" debug tool to immediatelty complete construction of the selected objects.
+* Fix detecting current drought state condition.
 
 # v0.12 (10 Jan 2024)
 * Experimental version for Update 5.